refactor: 1. modifed retouer to create Outlet. 2. created profileEdit file.
