+++ src/pages/ProfileEdit/index.tsx
@@ -1,0 +1,3 @@
+import ProfileEdit from "./profileEditStyle";
+
+export default ProfileEdit;
